Remove reviews from sign-up
